% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr.R
\name{ya_cpr}
\alias{ya_cpr}
\title{CPR}
\usage{
ya_cpr(base_df, calculate_df, customer_id_col, product_col)
}
\arguments{
\item{base_df}{data.frame 基期对照数据，需包含客户 ID（如结算代），以及产品}

\item{calculate_df}{data.frame 需要计算客户数据，需包含列同 \code{base_df}}

\item{customer_id_col}{客户区分列}

\item{product_col}{产品列}
}
\value{
data.frame 返回客户分类结果
}
\description{
计算 CPR
}
\details{
该函数用于根据客户和产品信息将客户分类为新客户、老客户新业务和老客户老业务。
}
