% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth-rate.R
\name{ya_growth_rate}
\alias{ya_growth_rate}
\title{增长率}
\usage{
ya_growth_rate(current_period, base_period)
}
\arguments{
\item{current_period}{当期值，一个非负数值向量。}

\item{base_period}{基期值，一个非负数值向量。}
}
\value{
一个数值向量，表示增长率。特殊情况处理如下：
\itemize{
\item 如果当期为零或 NA，且基期大于零，返回 -1
\item 如果当期大于零，且基期为零或 NA，返回 1
\item 如果当期和基期都不为零或 NA，返回当期除以基期减一
\item 如果当期和基期都为零或 NA，返回 NA
}
}
\description{
依据两个非负数值向量计算增长率，如果当期或基期为零或 NA，函数会返回特定值。适
用于同比、环比增长率计算。
}
