% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_per_customer.R
\name{ya_income_per_customer}
\alias{ya_income_per_customer}
\title{客户单产}
\usage{
ya_income_per_customer(data, periods, income_col, customer_id_col, ...)
}
\arguments{
\item{data}{数据框。}

\item{periods}{期数。}

\item{income_col}{收入列名。}

\item{customer_id_col}{用于区分客户计数的列名。}

\item{...}{分组列名。}
}
\value{
一个数据框，含分组列、客户单产。
}
\description{
根据指定的分组列，对数据进行分组，并计算每个分组的客户单产。
}
