% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_scientific_notation.R
\name{ya_parse_sci_str}
\alias{ya_parse_sci_str}
\title{解析科学计数法字符串为数值字符串}
\usage{
ya_parse_sci_str(x)
}
\arguments{
\item{x}{包含待解析字符串的字符向量。}
}
\value{
一个字符向量，包含解析后数值字符串，或解析失败的原字符串。
}
\description{
该函数接受一个字符串向量，尝试将其解析为数值，并返回解析后的值作为字符串。如
果解析失败，则返回原字符串。
}
