% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{yeahr_rename}
\alias{yeahr_rename}
\title{重命名文件}
\usage{
yeahr_rename(input_dir, output_dir, rename_vector)
}
\arguments{
\item{input_dir}{字符串，输入目录的路径。}

\item{output_dir}{字符串，输出目录的路径。}

\item{rename_vector}{命名向量，向量名为匹配的文本内容，向量值为重命名后的文件
名。}
}
\value{
无返回值。
}
\description{
根据提供的命名向量，重命名文件夹中文件。向量名为要匹配的文本内容，
向量值为重命名后的文件名。重命名后文件名，不需要包含文件格式后缀。
}
