% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpr.R
\name{ya_cpr}
\alias{ya_cpr}
\title{CPR}
\usage{
ya_cpr(base_df, calculate_df, customer_id_col, product_col)
}
\arguments{
\item{base_df}{数据框。基期数据，上年度有营收客户及产品清单。必需包含产品及客
户 ID，如结算代码。}

\item{calculate_df}{数据框。需要计算客户数据，必要列同 \code{base_df}，且需按必要
列分组求和。}

\item{customer_id_col}{客户区分列。}

\item{product_col}{产品列。}
}
\value{
数据框，含 \code{customer_id_col}，\code{product_col} 及 CPR 计算结果。
}
\description{
计算 CPR
}
\details{
该函数用于根据客户和产品信息将客户分类为新客户、老客户新业务和老客户老业务。
}
